change users.js to users.ts
