+++ package.json
@@ -36,7 +36,8 @@
   "dependencies": {
     "axios": "^1.5.0",
     "mainflux-sdk": "^0.14.38",
-    "node-fetch": "^3.3.2"
+    "node-fetch": "^3.3.2",
+    "typescript": "^5.4.3"
   },
   "devDependencies": {
     "esmify": "^2.1.1",
